pyproject.toml
[tool.poetry]
name = "commspt-bot-avilla"
version = "0.1.0"
description = ""
authors = ["SerinaNya <34389622+SerinaNya@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
avilla-core = "^1.0.0a24"
avilla-console = "^0.2.0"
avilla-elizabeth = "^1.0.0a26"
pydantic = "^2.5.3"
loguru = "^0.7.2"
arclet-alconna = {extras = ["full"], version = "^1.7.42"}
pygtrie = "^2.5.0"
graia-saya = "^0.0.19"
richuru = "^0.1.1"
pyyaml = "^6.0.1"
motor = "^3.3.2"
httpx = {extras = ["http2", "socks"], version = "^0.26.0"}
email-validator = "^2.1.0.post1"
arclet-alconna-avilla = "^0.17.3"


[[tool.poetry.source]]
name = "mirrors"
url = "https://mirror.sjtu.edu.cn/pypi/web/simple"
priority = "default"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
